Correct orm objects

+++ src/api/users.py
@@ -1,9 +1,12 @@
 from typing import List
 from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordRequestForm
+
 from src.models.schemas.user.user_request import UserRequest
 from src.models.schemas.user.user_response import UserResponse
 from src.services.users import UsersService
-
+from src.models.schemas.utils.jwt_token import JwtToken
+from src.services.users import get_current_user_role, get_current_user_id
 
 router = APIRouter(
     prefix='/users',
@@ -11,16 +14,28 @@ router = APIRouter(
 )
 
 
-@router.get('/all', response_model=List[UserResponse], name="Получить все резервуары")
-def get(user_service: UsersService = Depends()):
+@router.post('/authorize', response_model=JwtToken, name="Авторизация")
+def authorize(auth_schema: OAuth2PasswordRequestForm = Depends(), users_service: UsersService = Depends()) -> JwtToken:
+    result = users_service.authorize(auth_schema.username, auth_schema.password)
+    if not result:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Не авторизован.')
+    return result
+
+
+@router.get('/all', response_model=List[UserResponse], name="Получить всех пользователей")
+def get(user_service: UsersService = Depends(), user_role: str = Depends(get_current_user_role)):
     """
-    Получить все резервуары. Более подробное описание.
+    Получить всех пользователей. Более подробное описание.
     """
+    if not user_role == 'admin':
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='Недостаточно прав.')
     return user_service.all()
 
 
 @router.get('/get/{user_id}', response_model=UserResponse, name="Получить одну категорию")
-def get(user_id: int, users_service: UsersService = Depends()):
+def get(user_id: int, users_service: UsersService = Depends(), user_role: str = Depends(get_current_user_role)):
+    if not user_role == 'admin':
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='Недостаточно прав.')
     return get_with_check(user_id, users_service)
 
 
@@ -32,17 +47,33 @@ def get_with_check(user_id: int, users_service: UsersService):
 
 
 @router.post('/', response_model=UserResponse, status_code=status.HTTP_201_CREATED, name="Добавить категорию")
-def add(user_schema: UserRequest, users_service: UsersService = Depends()):
-    return users_service.add(user_schema)
+def add(
+        user_schema: UserRequest, users_service: UsersService = Depends(),
+        called_user_role: str = Depends(get_current_user_role),
+        called_user_id: int = Depends(get_current_user_id)
+        ):
+    if not called_user_role == 'admin':
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='Недостаточно прав.')
+    return users_service.add(user_schema, called_user_id)
 
 
 @router.put('/{user_id}', response_model=UserResponse, name="Обновить информацию о категории")
-def put(user_id: int, user_schema: UserRequest, users_service: UsersService = Depends()):
+def put(
+        user_id: int,
+        user_schema: UserRequest,
+        users_service: UsersService = Depends(),
+        called_user_role: str = Depends(get_current_user_role),
+        called_user_id: int = Depends(get_current_user_id)
+        ):
+    if not called_user_role == 'admin':
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='Недостаточно прав.')
     get_with_check(user_id, users_service)
-    return users_service.update(user_id, user_schema)
+    return users_service.update(user_id, user_schema, called_user_id)
 
 
 @router.delete('/{user_id}', status_code=status.HTTP_204_NO_CONTENT, name="Удалить категорию")
-def delete(user_id: int, users_service: UsersService = Depends()):
+def delete(user_id: int, users_service: UsersService = Depends(), user_role: str = Depends(get_current_user_role)):
+    if not user_role == 'admin':
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='Недостаточно прав.')
     get_with_check(user_id, users_service)
     return users_service.delete(user_id)
